% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitfunctions.R
\name{ilogit}
\alias{ilogit}
\title{Convert logits to probabilities}
\usage{
ilogit(log_odds)
}
\arguments{
\item{log_odds}{A numeric vector of positive or negative real numbers that
represent log odds, also known as logits.}
}
\value{
A vector of probabilities, which are values between 0.0 and 1.0.
}
\description{
This implements the inverse logit function. This function is also the
cumulative distribution function of the logistic distributions, and so is
available using the \code{\link[stats:Logistic]{stats::plogis()}} function.
}
\examples{
x <- rnorm(5)
ilogit(x)
}
