% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitfunctions.R
\name{logit}
\alias{logit}
\title{Convert probabilities to log odds}
\usage{
logit(p)
}
\arguments{
\item{p}{A numeric vector of probabilities, which are values between 0.0
and 1.0.}
}
\value{
A numeric vector of logits, which are positive or negative real numbers.
}
\description{
Log odds, also known as logits, are the logarithms,
usually to the base of the natural logarithms, of odds.
}
\examples{
p <- c(0.1, 0.25, 0.5, 0.9)
logit(p)
}
