% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitfunctions.R
\name{odds_to_probs}
\alias{odds_to_probs}
\title{Convert odds to probabilities}
\usage{
odds_to_probs(odds)
}
\arguments{
\item{odds}{A numeric vector of non-negative values, representing odds.}
}
\value{
A numeric vector of probabilities.
}
\description{
This is the inverse of the function that calculates odds from probabilities.
For any given value of odds, the corresponding probability \code{p} where
\code{odds = p/(1-p)} is returned.
}
\examples{
odds <- c(1, 3, 9)
odds_to_probs(odds)
}
