% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{probabilities}
\alias{probabilities}
\title{A tibble of probabilities}
\format{
A data frame with 5 rows and 1 variable:
\describe{
\item{p}{Important probabilities, each being a number between 0.0 and 1.0.}
}
}
\description{
A tibble data frame of very important probabilities.
}
