% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitfunctions.R
\name{probs_to_odds}
\alias{probs_to_odds}
\title{Convert probabilities to odds}
\usage{
probs_to_odds(p)
}
\arguments{
\item{p}{A numeric vector of probabilities, which are values between 0.0
and 1.0.}
}
\value{
A numeric vector of odds, which are values between zero and infinity.
}
\description{
Odds are an alternative means to quantify probabilities of events.
If the probability of an event has a value \code{p}, the odds corresponding
to \code{p} is \code{p/(1-p)}.
}
\examples{
p <- c(0.25, 0.5, 0.75)
probs_to_odds(p)
}
