% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdemopkg-package.R
\docType{package}
\name{rdemopkg-package}
\alias{rdemopkg}
\alias{rdemopkg-package}
\title{rdemopkg: Functions for Calculating Log Odds and Related Quantities}
\description{
Functions for calculating the odds, logarithm of odds, and their inverses. Logarithms of odds are also known as logits.
}
\author{
\strong{Maintainer}: Mark Andrews \email{mark.andrews@ntu.ac.uk}

}
\keyword{internal}
